Refactor UBC IP filter to a separate package. Add two simple tests.
